pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "efficient_tutor_backend"
version = "0.0.3"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "asyncpg>=0.30.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "dnspython==2.8.0",
    "dotenv==0.9.9",
    "email-validator==2.3.0",
    "fastapi[all]>=0.119.0",
    "flask==3.1.2",
    "flask-cors==6.0.1",
    "google-api-core==2.25.1",
    "google-api-python-client==2.181.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "googleapis-common-protos==1.70.0",
    "gunicorn==23.0.0",
    "httplib2==0.30.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "oauthlib==3.3.1",
    "packaging==25.0",
    "passlib[bcrypt]>=1.7.4",
    "pluggy==1.6.0",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pydantic-settings>=2.11.0",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "pyperclip==1.9.0",
    "pytest==8.4.2",
    "pytest-dotenv==0.5.2",
    "python-dotenv==1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "sqlalchemy>=2.0.44",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "werkzeug==3.1.3",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
pythonpath = [
    ".",
    "src",
]
